+++ internal/framework/service/rulesets/resource.go
@@ -352,8 +352,10 @@ func toRulesetResourceModel(ctx context.Context, zoneID, accountID basetypes.Str
 				ServerSideExcludes:      flatteners.Bool(ruleResponse.ActionParameters.ServerSideExcludes),
 				StatusCode:              flatteners.Int64(int64(ruleResponse.ActionParameters.StatusCode)),
 				SXG:                     flatteners.Bool(ruleResponse.ActionParameters.SXG),
+				OriginCacheControl:      flatteners.Bool(ruleResponse.ActionParameters.OriginCacheControl),
 				OriginErrorPagePassthru: flatteners.Bool(ruleResponse.ActionParameters.OriginErrorPagePassthru),
 				RespectStrongEtags:      flatteners.Bool(ruleResponse.ActionParameters.RespectStrongETags),
+				ReadTimeout:             flatteners.Int64(int64(cloudflare.Uint(ruleResponse.ActionParameters.ReadTimeout))),
 				Version:                 flatteners.String(cloudflare.String(ruleResponse.ActionParameters.Version)),
 			})
 
@@ -900,6 +902,10 @@ func (r *RulesModel) toRulesetRule(ctx context.Context) cloudflare.RulesetRule {
 			rr.ActionParameters.SXG = cloudflare.BoolPtr(ap.SXG.ValueBool())
 		}
 
+		if !ap.ReadTimeout.IsNull() {
+			rr.ActionParameters.ReadTimeout = cloudflare.UintPtr(uint(ap.ReadTimeout.ValueInt64()))
+		}
+
 		if !ap.Polish.IsNull() {
 			polish, _ := cloudflare.PolishFromString(ap.Polish.ValueString())
 			rr.ActionParameters.Polish = polish
@@ -917,6 +923,10 @@ func (r *RulesModel) toRulesetRule(ctx context.Context) cloudflare.RulesetRule {
 
 		if !ap.OriginErrorPagePassthru.IsNull() {
 			rr.ActionParameters.OriginErrorPagePassthru = cloudflare.BoolPtr(ap.OriginErrorPagePassthru.ValueBool())
+		}
+
+		if !ap.OriginCacheControl.IsNull() {
+			rr.ActionParameters.OriginCacheControl = cloudflare.BoolPtr(ap.OriginCacheControl.ValueBool())
 		}
 
 		if !ap.RespectStrongEtags.IsNull() {