+++ internal/sdkv2provider/resource_cloudflare_record.go
@@ -305,9 +305,11 @@ func resourceCloudflareRecordUpdate(ctx context.Context, d *schema.ResourceData,
 		updateRecord.TTL = ttl.(int)
 	}
 
-	// if comment, ok := d.GetOk("comment"); ok {
-	// 	updateRecord.Comment = comment.(string)
-	// }
+
+	comment, commentOk := d.GetOkExists("comment")
+	if commentOk {
+		updateRecord.Comment = cloudflare.StringPtr(comment.(string))
+	}
 
 	tags := []string{}
 	for _, tag := range d.Get("tags").(*schema.Set).List() {